fixed progress bar bug when downloading >100 cases
Previously the progress bar would not update if the user downloads > 100 cases because the incremental progress (which is less than 1%) was rounded off to the nearest integer, resulting in 0 incremental progress being signalled to the progress bar. Now instead of the incremental progress being rounded off, we round off the cumulative progress.